MobileApp/requirements.txt
Kivy==2.1.0
PyYAML==6.0
pipreqs==0.4.11
pathlib==1.0.1